feat: remake non-sign in home and discover
