+++ lib/i18n/auth/auth_en.i18n.json
@@ -7,5 +7,10 @@
     "lets_get_you_in": "Let's get you in!",
     "get_started_description": "Sign up or log in to create, collaborate & celebrate! #makelemonade",
     "get_started": "Get Started",
-    "signIn": "Sign In"
+    "signIn": "Sign In",
+    "getStarted": "Get Started",
+    "getStartedDescription": "Register for events, subscribe to communities and manage events you are going to ",
+    "signInOrCreateAccount": "Sign In or Create Account",
+    "explore": "Explore"
+    
 }
